[refactor] COPY 버튼 구현 및 페이지 레이아웃 리펙토링
